+++ src/main/webapp/resources/js/pages/projects/remote/index.js
@@ -1,0 +1,5 @@
+import React from "react";
+import { render } from "react-dom";
+import { ProjectSynchronizationSettings } from "./ProjectSynchronizationSettings";
+
+render(<ProjectSynchronizationSettings />, document.querySelector("#project-remote-root"));